+++ src/widgets/Input/index.tsx
@@ -1,5 +1,5 @@
 import cs from 'classnames';
-import { ChangeEvent, ForwardedRef, forwardRef, useState } from 'react';
+import { ChangeEvent, useRef, useState } from 'react';
 import styles from './index.module.css';
 import { InputProps } from './Input.props';
 
@@ -8,22 +8,32 @@ import { InputProps } from './Input.props';
  * @param {string} textToAction - (byDefault='') text to do action with
  * @returns {any}
  */
-/**TODO: оформить ref - выставить фокус у элемента формы */
+
 /**
  * Util Input component 
  * @param {requestCallback} onSend - callback to do request after submit
- * @param {string} iconClassName additional CSS class for icon (used for setting icon image)
  * @param {Object} ref (optional, byDefault=null) - to be focused after submit
+ * @param {string} placeholder - 
  * @returns {component} Input 
  * 
 */
 
-const Input = forwardRef(function Input(props: InputProps, ref: ForwardedRef<HTMLInputElement>) {
-	const {className, iconClassName, placeholder} = props as InputProps;
+export default function Input(props: InputProps) {
+	const { placeholder } = props as InputProps;
 	const [text, setText] = useState<string>('');
+	const inputRef = useRef<HTMLInputElement>(null);
+
+	const clearAndFocusAfterSubmit = () => {
+		if(inputRef?.current) {
+			inputRef.current.value = '';
+			inputRef.current.focus();
+		}
+	};
+
 	const onClick = () => {
 		if (typeof text == 'string') {
 			(props as InputProps).onSend(text);
+			clearAndFocusAfterSubmit();
 		}
 	};
 
@@ -34,10 +44,9 @@ const Input = forwardRef(function Input(props: InputProps, ref: ForwardedRef<HTM
 
 	return (
 		<div className={cs(styles['input-container'])}>
-			<input value={text} onChange={onChange} placeholder={placeholder} name='textToAction' ref={ref} className={cs(styles['input'], className)} />
-			<img  className={cs(styles['icon'], iconClassName)} src='/arrow-input-icon.svg' alt='Стрелка отправить' onClick={onClick}/>
+			<input ref={inputRef} value={text} onChange={onChange} placeholder={placeholder} name='textToAction' className={cs(styles['input'])} />
+			<img  className={styles['icon']} src='/arrow-input-icon.svg' alt='Стрелка отправить' onClick={onClick}/>
 		</div>
 	);
-});
+};
 
-export default Input;